mobiXAI/requirements.txt
Flask>=2.0.1
Werkzeug>=2.0.1
Jinja2>=3.0.1
xgboost==2.0.2
lime==0.2.0.1
pandas
numpy
scikit-learn
openai
anthropic
google-generativeai
cohere
gunicorn==20.1.0
requests==2.25.1